backend/requirements.txt
fastapi
uvicorn
motor
pydantic
pydantic-settings
email-validator
python-jose[cryptography]
passlib[bcrypt]
python-multipart
python-dotenv
pandas
openpyxl
fpdf2
python-docx
